@@ -1,44 +1,56 @@
 <?php
 
-namespace Atomescrochus\EPF\Commands;
+namespace Appwapp\EPF\Commands;
 
-use Atomescrochus\EPF\EPFCrawler;
-use Atomescrochus\EPF\Exceptions\MissingCommandOptions;
-use Atomescrochus\EPF\Traits\FeedCredentials;
-use Atomescrochus\EPF\Traits\FileStorage;
-use GuzzleHttp\Client;
-use Illuminate\Console\Command;
-use Illuminate\Support\Facades\Storage;
-use Symfony\Component\Console\Helper\ProgressBar;
+use Appwapp\EPF\EPFCrawler;
+use Appwapp\EPF\Exceptions\ModelNotFoundException;
+use Appwapp\EPF\Jobs\DownloadJob;
+use Appwapp\EPF\Traits\FeedCredentials;
+use Appwapp\EPF\Traits\FileStorage;
+use Illuminate\Support\Str;
 
-class EPFDownloader extends Command
+class EPFDownloader extends EPFCommand
 {
-    use FileStorage;
-    use FeedCredentials;
+    use FeedCredentials, FileStorage;
 
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'epf:download';
+    protected $signature = 'epf:download
+        {--type= : the type of import, either "full", or "incremental"}
+        {--group= : the group of import, either "itunes", "match", "popularity" or "pricing"}
+        {--skip-confirm : skip the confirmation prompt}
+        {--chain-jobs : chain the next jobs after the download}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'This will help download the EPF files.';
+    protected $description = 'Download the EPF export files to local storage.';
 
-    private $credentials;
+    /**
+     * The EPF crawler.
+     *
+     * @var \Appwapp\EPF\EPFCrawler
+     */
+    private EPFCrawler $epf;
+
+    /**
+     * The paths.
+     *
+     * @var object
+     */
     private $paths;
-    private $epf;
-    private $type;
-    private $group;
-    private $md5ChecksFailed;
-    private $downloadProgressBar;
-    private $fullPathToFiles;
-    private $variableFolders;
+
+    /**
+     * The credentials.
+     *
+     * @var object
+     */
+    private $credentials;
 
     /**
      * Create a new command instance.
@@ -50,9 +62,7 @@ class EPFDownloader extends Command
         parent::__construct();
 
         $this->credentials = $this->getCredentials();
-        $this->paths = $this->getEPFFilesPaths();
-
-        $this->md5ChecksFailed = false;
+        $this->paths       = $this->getEPFFilesPaths();
     }
 
     /**
@@ -62,131 +72,59 @@ class EPFDownloader extends Command
      */
     public function handle()
     {
-        dd("yolo");
         $this->line("");
-        $this->line("👋. Welcome to the Apple EPF downloader! 👋");
+        $this->line("👋 Welcome to the Apple EPF downloader! 👋");
 
-        $this->type = $this->choice('What is the type of files you want to download?', ['full', 'incremental'], 0);
-        $this->group = $this->choice('What is the group of files you want to download?', ['itunes', 'match', 'popularity', 'pricing'], 0);
+        // Get the group and type
+        $this->gatherUserInput();
 
-        $this->variableFolders = "{$this->group}/{$this->type}";
-
-        if ($this->confirm("Are you ready to launch the downloads?")) {
+        if ($this->option('skip-confirm') || $this->confirm("Are you ready to launch the downloads?")) {
             $this->epf = new EPFCrawler();
             $this->startTasks();
+
+            $this->line('');
             $this->comment("We're done downloading!");
         }
     }
 
-    private function startTasks()
+    /**
+     * Start the download tasks.
+     *
+     * @return void
+     */
+    private function startTasks(): void
     {
         $this->line("");
         $this->line("We're starting to download the '{$this->group}|{$this->type}' group of files. Depending on your connection, this might take a long time!");
 
-        $links = $this->epf->links->get($this->type)->filter(function ($link) {
-            return str_contains($link, $this->group);
-        });
-        // $links = collect(["https://feeds.itunes.apple.com/feeds/epf/v3/full/current/incremental/current/match20170323.tbz"]); // debug only
+        $links      = $this->epf->links->get($this->type)->get($this->group);
         $countLinks = count($links);
-        $this->info("There is a total of {$countLinks} files to download.");
+           
+        $this->info("There is a total of $countLinks files to download.");
 
         $links->each(function ($link) {
+            $this->line("");
             $filename = basename($link);
 
-            $this->line("");
-            $this->info("Starting download of {$filename}");
+            // Fetch the model based on the file name and group
+            $model = 'Appwapp\EPF\Models\\'. Str::studly($this->group)  .'\\' . Str::studly(Str::replace('.tbz', '', $filename));
 
-            $this->download($link);
-            $this->md5Check($link);
+            // Remove any number, in case of multiple files for same model
+            $model = preg_replace('/[0-9]*/', '', $model);
 
-            $this->info("Finished download of {$filename}");
-            $this->line("");
+            // If model does not exist, throw an error
+            if (! class_exists($model)) {
+                throw new ModelNotFoundException("Model '$model' does not exists. Make sure 'apple-epf-laravel' is up to date.");
+            }
+
+            // Check if we need to skip that file/model
+            if (! in_array($model, config('apple-epf.included_models'))) {
+                $this->comment("Skipped download of {$filename}");
+                return;
+            }
+
+            $this->line("Dispatching download of {$filename}...");
+            DownloadJob::dispatch($link, $this->group, $this->type)->onQueue(config('apple-epf.queue'));
         });
-    }
-
-    private function download($link)
-    {
-        $this->line("");
-        $linkMD5 = $link.".md5";
-        $filename = basename($link);
-        $filenameMD5 = basename($linkMD5);
-        
-        // make sure file exists
-        // todo: this action overwrites the file if it exists, should be better...
-        $file = Storage::put("{$this->paths->get('storage')->archive}/{$this->variableFolders}/{$filename}", "");
-        $fileMd5 = Storage::put("{$this->paths->get('storage')->archive}/{$this->variableFolders}/{$filenameMD5}", "");
-        
-        $client = new Client();
-        $fileResponse = $client->request('GET', $link, [
-            'auth' => [$this->credentials->login, $this->credentials->password],
-            'sink' => "{$this->paths->get('system')->archive}/{$this->variableFolders}/{$filename}",
-            'progress' => function ($downloadTotal, $downloadedBytes, $uploadTotal, $uploadedBytes) {
-                if ($this->downloadProgressBar == null && $downloadTotal != 0) {
-                    // no progress bar and download started
-                        
-                    $this->bytesInPreviousIteration = 0; //
-                    $this->downloadProgressBar = new ProgressBar($this->output, $downloadTotal);
-                    $this->downloadProgressBar->setFormat('very_verbose');
-                } else if ($this->downloadProgressBar != null && $downloadTotal == $downloadedBytes) {
-                    // there is a progress bar and the download it finished
-                        
-                    $this->downloadProgressBar->finish();
-                    $this->downloadProgressBar->clear();
-                    $this->downloadProgressBar = null;
-                } else if ($this->downloadProgressBar != null) {
-                    // at this point, we're downloading and got a progress bar
-
-                    $this->downloadProgressBar->advance($downloadedBytes - $this->bytesInPreviousIteration);
-                    $this->bytesInPreviousIteration = $downloadedBytes;
-                }
-            },
-        ]);
-
-        $fileMd5Response = $client->request('GET', $linkMD5, [
-            'auth' => [$this->credentials->login, $this->credentials->password],
-            'sink' => "{$this->paths->get('system')->archive}/{$this->variableFolders}/{$filenameMD5}",
-            'progress' => function ($downloadTotal, $downloadedBytes, $uploadTotal, $uploadedBytes) {
-                if ($this->downloadProgressBar == null && $downloadTotal != 0) {
-                    // no progress bar and download started
-                        
-                    $this->bytesInPreviousIteration = 0; //
-                    $this->downloadProgressBar = new ProgressBar($this->output, $downloadTotal);
-                    $this->downloadProgressBar->setFormat('very_verbose');
-                } else if ($this->downloadProgressBar != null && $downloadTotal == $downloadedBytes) {
-                    // there is a progress bar and the download it finished
-                        
-                    $this->downloadProgressBar->finish();
-                    $this->downloadProgressBar->clear();
-                    $this->downloadProgressBar = null;
-                } else if ($this->downloadProgressBar != null) {
-                    // at this point, we're downloading and got a progress bar
-
-                    $this->downloadProgressBar->advance($downloadedBytes - $this->bytesInPreviousIteration);
-                    $this->bytesInPreviousIteration = $downloadedBytes;
-                }
-            },
-        ]);
-
-        $this->line("");
-    }
-
-    private function md5Check($file)
-    {
-
-        $filename = basename($file);
-        $md5Filename = $filename.".md5";
-        $fileLocation = "{$this->paths->get('system')->archive}/{$this->variableFolders}/{$filename}";
-        $md5FileLocation = "{$this->paths->get('storage')->archive}/{$this->variableFolders}/{$md5Filename}";
-
-        $md5File = md5_file($fileLocation);
-        $md5 = trim(substr(Storage::get($md5FileLocation), -33));
-
-        if ($md5File == $md5) {
-            $this->line("Checksum: ✅  passed!");
-            Storage::delete($md5FileLocation);
-            return true;
-        } else {
-            die("Checsum: ❌  failed! We'll have to quit, sorry!");
-        }
-    }
+    }    
 }
