@@ -1,39 +1,40 @@
 <?php
 
-namespace Atomescrochus\EPF\Commands;
+namespace Appwapp\EPF\Commands;
 
-use Alchemy\Zippy\Zippy;
-use Atomescrochus\EPF\Traits\FeedCredentials;
-use Atomescrochus\EPF\Traits\FileStorage;
-use Illuminate\Console\Command;
+use Appwapp\EPF\Jobs\ExtractJob;
+use Appwapp\EPF\Traits\FileStorage;
 use Illuminate\Support\Facades\Storage;
 
-class EPFExtractor extends Command
+class EPFExtractor extends EPFCommand
 {
     use FileStorage;
-    use FeedCredentials;
 
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'epf:extract';
+    protected $signature = 'epf:extract
+        {--type= : the type of import, either "full", or "incremental"}
+        {--group= : the group of import, either "itunes", "match", "popularity" or "pricing"}
+        {--file= : the file you want to extract, either "all" or the file path you want to extract}
+        {--skip-confirm : skip the confirmation prompt}
+        {--delete : delete the archive after extraction}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'This will help extract files downloaded from epf:download.';
+    protected $description = 'Extract the EPF files downloaded from epf:download.';
 
-    private $credentials;
+    /**
+     * The paths.
+     *
+     * @var object
+     */
     private $paths;
-
-    private $type;
-    private $group;
-    private $md5ChecksFailed;
-    private $variableFolders;
 
     /**
      * Create a new command instance.
@@ -44,7 +45,6 @@ class EPFExtractor extends Command
     {
         parent::__construct();
 
-        $this->credentials = $this->getCredentials();
         $this->paths = $this->getEPFFilesPaths();
     }
 
@@ -56,60 +56,55 @@ class EPFExtractor extends Command
     public function handle()
     {
         $this->line("");
-        $this->line("👋. Welcome to the Apple EPF extractor! 👋");
+        $this->line("👋 Welcome to the Apple EPF extractor! 👋");
 
-        $this->type = $this->choice('What is the type of files you want to extract?', ['full', 'incremental'], 0);
-        $this->group = $this->choice('What is the group of files you want to extract?', ['itunes', 'match', 'popularity', 'pricing'], 0);
+        // Get the group and type
+        $this->gatherUserInput();
 
-        $this->variableFolders = "{$this->group}/{$this->type}";
-
-        $listOfFiles = Storage::files($this->paths->get('storage')->archive."/{$this->variableFolders}");
+        // Get the list of downloaded archive files
+        $listOfFiles    = Storage::files($this->paths->get('storage')->archive."/{$this->variableFolders}");
         $extractChoices = array_merge(['all'], $listOfFiles);
-
-        if (count($extractChoices) == 1) {
-            die("Can't find files for those choices, did you download the files first?");
+        if (count($extractChoices) === 1) {
+            $this->error('Can\'t find files to extract, did you download the files first?');
+            exit;
         }
 
-        $toExtract = $this->choice('Which file do you want to extract?', $extractChoices, 0);
+        $toExtract = $this->option('file') ?? $this->choice('Which file do you want to extract?', $extractChoices, 0);
         $toExtract = $toExtract == "all" ? $listOfFiles : $toExtract;
 
-        if ($this->confirm("Are you ready to launch the extraction process?")) {
+        if ($this->option('skip-confirm') !== null || $this->confirm("Are you ready to launch the extraction process?")) {
             $this->startTasks($toExtract);
             $this->comment("We're done extracting!");
         }
     }
 
-    public function startTasks($toExtract)
+    /**
+     * Start the extraction tasks.
+     * 
+     * @param array $toExtract
+     *
+     * @return void
+     */
+    public function startTasks(array $toExtract)
     {
-        $this->checkForExtractionFolder();
-        $this->startExtractionProcess($toExtract);
-    }
+        // Create the extraction directory
+        Storage::makeDirectory($this->paths->get('storage')->extraction . "/{$this->variableFolders}");
 
-    public function checkForExtractionFolder()
-    {
-        Storage::makeDirectory($this->paths->get('storage')->extraction."/{$this->variableFolders}");
-    }
-
-    public function startExtractionProcess($toExtract)
-    {
         $toExtract = collect($toExtract);
 
         $toExtract = $toExtract->map(function ($file) {
-            return $this->paths->get('system')->storage.$file;
+            return $this->paths->get('system')->storage . $file;
         });
 
         $toExtract->each(function ($file) {
             $filename = basename($file);
-
-            $this->info("Extracting {$filename} ...");
-
-            $extractor = Zippy::load();
-            $archive = $extractor->open($file, '.tar.bz2');
-            $archive->extract($this->paths->get('system')->extraction."/{$this->variableFolders}");
-
-            if ($this->confirm("Extraction of {$filename} completed, do you wish to delete the archive?")) {
-                Storage::delete($this->paths->get('storage')->archive."/{$this->variableFolders}/{$filename}");
-            }
+            $this->line("Dispatching extraction of $filename...");
+            ExtractJob::dispatch(
+                $file,
+                $this->group,
+                $this->type,
+                $this->option('delete') !== null || $this->confirm("Do you wish to delete the archive?")
+            )->onQueue(config('apple-epf.queue'));
         });
     }
 }
