@@ -1,56 +1,127 @@
 <?php
 
-namespace Atomescrochus\EPF;
+namespace Appwapp\EPF;
 
-use Atomescrochus\EPF\Traits\FeedCredentials;
-use Atomescrochus\EPF\Traits\FileStorage;
-use Carbon\Carbon;
+use Appwapp\EPF\Traits\FeedCredentials;
+use Appwapp\EPF\Traits\FileStorage;
+use Illuminate\Support\Carbon;
 use GuzzleHttp\Client;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\DomCrawler\Crawler;
+use Illuminate\Support\Collection;
 
 class EPFCrawler
 {
-    use FeedCredentials;
-    use FileStorage;
-
-    private $credentials;
-    protected $currentIndexContent;
-    protected $currentCrawlerUrl;
-
-    protected $urlForFullImportFiles;
-    protected $urlForIncrementalImportFiles;
-
-    public $links;
-    public $fullImportTime;
-    public $incrementalImportTime;
+    use FeedCredentials, FileStorage;
 
     /**
-     * Create a new Skeleton Instance.
+     * The index of the current content.
+     *
+     * @var mixed
+     */
+    protected $currentIndexContent;
+
+    /**
+     * The URL of the current crawled content.
+     *
+     * @var string
+     */
+    protected string $currentCrawlerUrl;
+
+    /**
+     * URL for the full import.
+     * @var mixed
+     */
+    protected string $urlForFullImportFiles;
+
+    /**
+     * URL for the incremental import.
+     * @var string
+     */
+    protected string $urlForIncrementalImportFiles;
+
+    /**
+     * All the links to crawl for.
+     *
+     * @var Collection
+     */
+    public Collection $links;
+
+    /**
+     * The time of the full import.
+     *
+     * @var Carbon
+     */
+    public Carbon $fullImportTime;
+
+    /**
+     * The time of the incremental import.
+     *
+     * @var Carbon
+     */
+    public Carbon $incrementalImportTime;
+
+    /**
+     * Construct a new instance.
      */
     public function __construct()
     {
-        $this->credentials = $this->getCredentials();
-        $this->paths = $this->getEPFFilesPaths();
-
-        $this->urlForFullImportFiles = "https://feeds.itunes.apple.com/feeds/epf/v3/full/current/";
-        $this->urlForIncrementalImportFiles = "https://feeds.itunes.apple.com/feeds/epf/v3/full/current/incremental/current/";
+        $this->urlForFullImportFiles        = "https://feeds.itunes.apple.com/feeds/epf/v5/current/";
+        $this->urlForIncrementalImportFiles = "https://feeds.itunes.apple.com/feeds/epf/v5/current/incremental/current/";
 
         $this->links = collect([
-            'full' => $this->getFullImportListOfFiles(),
-            'incremental' => $this->getIncrementalImportListOfFiles(),
+            'full'        => collect($this->getFullImportListOfTypes()),
+            'incremental' => collect($this->getIncrementalImportListOfTypes()),
         ]);
 
         $this->currentIndexContent = "";
-        $this->credentials = "";
     }
 
-    private function getFullImportListOfFiles()
+    /**
+     * Gets a full import list per type.
+     * i.e. itunes, match, popularity and pricing
+     *
+     * @return array
+     */
+    private function getFullImportListOfTypes(): array
     {
-        $this->crawlCurrentFolder(false);
+        $this->crawlFolder($this->urlForFullImportFiles);
 
+        $types   = [];
         $crawler = new Crawler($this->currentIndexContent, $this->urlForFullImportFiles);
-        $links = collect($crawler->filter('table > tr > td > a')->links());
+        $links   = collect($crawler->filter('table > tr > td > a')->links());
+        $links   = $links->reject(function ($link) {
+            return str_contains($link->getUri(), 'incremental');
+        });
+
+        // Get the links for each type
+        foreach ($links as $link) {
+            // Get the type name from link by removing the date
+            $type = preg_replace('/[0-9]+|\//', '', $link->getNode()->nodeValue);
+
+            // Get the links for that type
+            $types[$type] = $this->getFullImportListOfFiles($link->getUri());
+        }
+
+        // Get the import time
+        $this->fullImportTime = $this->getDateFromFilename($links->first()->getNode()->nodeValue);
+
+        return $types;
+    }
+
+    /**
+     * Gets a full import list of files from the EPF directory.
+     * 
+     * @param  string $url
+     *
+     * @return \Illuminate\Support\Collection
+     */
+    private function getFullImportListOfFiles(string $url): Collection
+    {
+        $this->crawlFolder($url);
+
+        $crawler = new Crawler($this->currentIndexContent, $url);
+        $links   = collect($crawler->filter('table > tr > td > a')->links());
 
         $links =  $links->reject(function ($link) {
             return str_contains($link->getUri(), 'incremental');
@@ -58,18 +129,51 @@ class EPFCrawler
             return str_contains($link->getUri(), '.md5');
         })->map(function ($link) {
             return $link->getUri();
-        });
-
-        $this->fullImportTime = $this->getDateFromFilename($links->first());
+        });        
 
         return $links;
     }
 
-    private function getIncrementalImportListOfFiles()
+    /**
+     * Gets a incremental import list per type.
+     * i.e. itunes, match, popularity and pricing
+     *
+     * @return array
+     */
+    private function getIncrementalImportListOfTypes(): array
     {
-        $this->crawlCurrentFolder(true);
+        $this->crawlFolder($this->urlForIncrementalImportFiles);
 
+        $types   = [];
         $crawler = new Crawler($this->currentIndexContent, $this->urlForIncrementalImportFiles);
+        $links   = collect($crawler->filter('table > tr > td > a')->links());
+
+        foreach ($links as $link) {
+            // Get the type name from link by removing the date
+            $type = preg_replace('/[0-9]+|\//', '', $link->getNode()->nodeValue);
+
+            // Get the links for that type
+            $types[$type] = $this->getIncrementalImportListOfFiles($link->getUri());
+        }
+
+        // Get the import time
+        $this->incrementalImportTime = $this->getDateFromFilename($links->first()->getNode()->nodeValue);
+
+        return $types;
+    }
+
+    /**
+     * Gets the incremental import list of files from the EPF directory.
+     *
+     * @param  string $url
+     *
+     * @return \Illuminate\Support\Collection
+     */
+    private function getIncrementalImportListOfFiles(string $url): Collection
+    {
+        $this->crawlFolder($url);
+
+        $crawler = new Crawler($this->currentIndexContent, $url);
         $links = collect($crawler->filter('table > tr > td > a')->links());
 
         $links = $links->map(function ($link) {
@@ -78,29 +182,44 @@ class EPFCrawler
             return str_contains($link, '.md5');
         });
 
-        $this->incrementalImportTime = $this->getDateFromFilename($links->first());
-
         return $links;
     }
 
-    private function getDateFromFilename($link)
+    /**
+     * Gets the date from the import filename.
+     *
+     * @param  string $link
+     *
+     * @return Carbon
+     */
+    private function getDateFromFilename(string $link): Carbon
     {
         $filename = basename($link, ".tbz");
         $time = str_replace("itunes", "", $filename);
         return Carbon::parse($time);
     }
 
-    private function crawlCurrentFolder($incremental = false)
+    /**
+     * Crawl the folder.
+     *
+     * @param  string  $url  The remote url to crawl
+     *
+     * @return void
+     */
+    private function crawlFolder(string $url): void
     {
-        $currentCrawlerUrl = $incremental ? $this->urlForIncrementalImportFiles : $this->urlForFullImportFiles;
+        $paths       = $this->getEPFFilesPaths();
+        $credentials = $this->getCredentials();
 
-        Storage::put($this->paths->get('storage')->epf_folder."/index.html", "");
+        Storage::put($paths->get('storage')->epf_folder."/index.html", "");
 
         $client = new Client();
-        
-        $client->request('GET', $currentCrawlerUrl, ['auth' => [$this->credentials->login, $this->credentials->password],'sink' => $this->paths->get('system')->epf_folder."/index.html"]);
+        $client->request('GET', $url, [
+            'auth' => [$credentials->login, $credentials->password],
+            'sink' => $paths->get('system')->epf_folder . '/index.html'
+        ]);
 
-        $filePath = $this->paths->get('storage')->epf_folder."/index.html";
+        $filePath = $paths->get('storage')->epf_folder . '/index.html';
         
         $this->currentIndexContent = Storage::get($filePath);
         Storage::delete($filePath);
