@@ -1,14 +1,18 @@
 <?php
 
-namespace Atomescrochus\EPF\Traits;
+namespace Appwapp\EPF\Traits;
 
 trait FeedCredentials
 {
-    
-    public function getCredentials()
+    /**
+     * Get the Apple EPF user credentials from configuration.
+     *
+     * @return object
+     */
+    public function getCredentials(): object
     {
-        return  (object) [
-            'login' => config('apple-epf.user_id'),
+        return (object) [
+            'login'    => config('apple-epf.user_id'),
             'password' => config('apple-epf.password'),
         ];
     }
